+++ dashboard/src/components/CheckList.js
@@ -11,6 +11,7 @@ import Checkbox from "@material-ui/core/Checkbox";
 import Row from "/components/CheckRow";
 
 import Loader from "/components/util/Loader";
+import Pagination from "/components/partials/Pagination";
 
 class CheckList extends React.Component {
   static propTypes = {
@@ -20,30 +21,40 @@ class CheckList extends React.Component {
       }),
     }),
     loading: PropTypes.bool,
+    onChangeQuery: PropTypes.func.isRequired,
+    limit: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    offset: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   };
 
   static defaultProps = {
     environment: null,
     loading: false,
+    limit: undefined,
+    offset: undefined,
   };
 
   static fragments = {
     environment: gql`
       fragment CheckList_environment on Environment {
-        checks(limit: $limit) {
+        checks(limit: $limit, offset: $offset) {
           nodes {
             id
             ...CheckRow_check
+          }
+
+          pageInfo {
+            ...Pagination_pageInfo
           }
         }
       }
 
       ${Row.fragments.check}
+      ${Pagination.fragments.pageInfo}
     `,
   };
 
   render() {
-    const { environment, loading } = this.props;
+    const { environment, loading, limit, offset, onChangeQuery } = this.props;
     const checks = environment ? environment.checks.nodes : [];
 
     return (
@@ -64,6 +75,13 @@ class CheckList extends React.Component {
             {checks.map(node => <Row key={node.id} check={node} />)}
           </TableBody>
         </Table>
+
+        <Pagination
+          limit={limit}
+          offset={offset}
+          pageInfo={environment && environment.checks.pageInfo}
+          onChangeQuery={onChangeQuery}
+        />
       </Loader>
     );
   }