+++ dashboard/src/components/views/EnvironmentView/EventsContent.js
@@ -23,6 +23,8 @@ class EventsContent extends React.Component {
     queryParams: PropTypes.shape({
       filter: PropTypes.string,
       order: PropTypes.string,
+      offset: PropTypes.string,
+      limit: PropTypes.string,
     }).isRequired,
     setQueryParams: PropTypes.func.isRequired,
   };
@@ -31,6 +33,8 @@ class EventsContent extends React.Component {
     query EnvironmentViewEventsContentQuery(
       $filter: String = "${defaultExpression}"
       $order: EventsListOrder = SEVERITY
+      $limit: Int,
+      $offset: Int,
       $environment: String!
       $organization: String!
     ) {
@@ -43,11 +47,15 @@ class EventsContent extends React.Component {
   `;
 
   render() {
+    const { queryParams, setQueryParams, match } = this.props;
+
+    const { filter, order, limit = "50", offset = "0" } = queryParams;
+
     return (
       <Query
         query={EventsContent.query}
         fetchPolicy="cache-and-network"
-        variables={{ ...this.props.match.params, ...this.props.queryParams }}
+        variables={{ ...match.params, filter, order, limit, offset }}
       >
         {({ data: { environment } = {}, loading, aborted, refetch }) => {
           if (!environment && !loading && !aborted) {
@@ -61,8 +69,8 @@ class EventsContent extends React.Component {
                   renderSearch={
                     <SearchBox
                       placeholder="Filter events…"
-                      initialValue={this.props.queryParams.filter}
-                      onSearch={filter => this.props.setQueryParams({ filter })}
+                      initialValue={filter}
+                      onSearch={value => setQueryParams({ filter: value })}
                     />
                   }
                   renderMenuItems={
@@ -75,7 +83,9 @@ class EventsContent extends React.Component {
                 />
               </Content>
               <EventsList
-                onChangeQuery={this.props.setQueryParams}
+                limit={limit}
+                offset={offset}
+                onChangeQuery={setQueryParams}
                 environment={environment}
                 loading={loading || aborted}
               />
@@ -87,4 +97,6 @@ class EventsContent extends React.Component {
   }
 }
 
-export default withQueryParams(["filter", "order"])(EventsContent);
+export default withQueryParams(["filter", "order", "offset", "limit"])(
+  EventsContent,
+);