+++ CHANGELOG.md
@@ -19,6 +19,7 @@ prior to assuming the existence of said check.
 - Added the format `wrapped-json` to sensuctl `configure`, `list` and `info`
 commands, which is compatible with `sensuctl create`.
 - Added debug event log with all event data.
+- Added environment variables to checks.
 
 ### Changed
 - Add logging around the Sensu event pipeline.
@@ -35,6 +36,7 @@ when creating resources with `sensuctl create`.
 - Use a default hostname if one cannot be retrieved.
 - Return an error from `sensuctl configure` when the configured organization
 or environment does not exist.
+- Remove an unnecessary parameter from sensuctl environment create.
 
 ### Fixed
 - Prevent panic when verifying if a metric event is silenced.